36_lesson_Таймер обратного отсчета (CountDownTimer) это класс который нам позваляет отсчитывать заданное время назад. На этом уроке я вам покажу как с ним работать. Используем как всегда Android Studio и язык программирования Kotlin.
