+++ .idea/misc.xml
@@ -3,7 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/react/AndroidStudioProjects/CristalBall/app/src/main/res/layout/activity_main.xml" value="0.38333333333333336" />
+        <entry key="..\:/Users/react/AndroidStudioProjects/CountDownTimer/app/src/main/res/layout/activity_main.xml" value="0.37395833333333334" />
       </map>
     </option>
   </component>
