+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.cristalball"
+        applicationId "com.example.countdowntimer"
         minSdk 23
         targetSdk 32
         versionCode 1
