--- app/src/main/java/com/example/cristalball/MainActivity.kt
@@ -1,25 +1,0 @@
-package com.example.cristalball
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.cristalball.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        binding.tvReto.setOnClickListener {
-            binding.tvReto.text = getReto()
-        }
-    }
-    private fun getReto(): String{
-        return resources.getStringArray(R.array.retos)[randomNumber()]
-    }
-    private fun randomNumber(): Int{
-        val size = resources.getStringArray(R.array.retos).size - 1
-        return (0..size).random()
-    }
-}
