--- app/src/main/res/values/arrays.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string-array name="retos">
-        <item>Жди денег</item>
-        <item>Осторожно</item>
-        <item>Всё будет хорошо</item>
-        <item>Потеряешь деньги</item>
-        <item>Любовь</item>
-        <item>Новое знакомство</item>
-    </string-array>
-</resources>
