+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Cristal Ball"
+rootProject.name = "CountDownTimer"
 include ':app'
